requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
distlib==0.3.6
filelock==3.12.2
Flask==2.3.2
Flask-Mobility==1.1.0
frozenlist==1.3.3
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.25.2
openai==0.27.8
packaging==23.1
pipenv==2023.6.26
Pygments==2.15.1
python-docx==0.8.11
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rich==13.4.2
safetensors==0.3.2
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.31.0
typing_extensions==4.7.1
urllib3==2.0.3
virtualenv==20.23.1
virtualenv-clone==0.5.7
Werkzeug==2.3.6
yarl==1.9.2
